requirements.txt

accelerate==0.22.0
diffusers==0.12.1
huggingface-hub==0.16.4
matplotlib==3.7.2
numpy==1.25.2
openai==0.27.9
pandas==2.0.3
Pillow==10.0.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
torch==2.0.1
torchmetrics==1.1.0
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.0
wandb==0.15.8
